Extract EPUB draft history helpers
